sessions/requirements.txt
uvicorn==0.29.0
fastapi==0.110.1
pytest==8.1.1
redis[hiredis]==5.0.3
python-dotenv==1.0.1
